added parsers, added tests

+++ src/main/java/com/agnieszka/piotrowska/weatherCheckApp/model/response/NearestInstallationResponse.java
@@ -7,7 +7,8 @@ import lombok.Data;
 
 @Data
 public class NearestInstallationResponse {
-    private Long id;
+
+    private int id;
     private Location location;
     private Address address;
     private Double elevation;
